refactor: rename the shared package to wahlchat-common
The name no longer ties it to the corpus. Today its contents are the
corpus contract plus Vertex credentials, but anything the Python
components come to share belongs there — the Firebase functions are
Python and could become a third consumer — and a corpus-specific name
would have pushed that toward a second shared package instead.

It also resolves the one thing that sat oddly before: vertex_credentials
serves llms.py for chat models as much as it serves embeddings, so it
never really belonged in something called wahlchat-corpus.

The rule that keeps it honest is unchanged and now stated in its README:
every dependency it declares must already be a direct dependency of both
consumers, or it lands in every image.
