+++ Makefile
@@ -129,7 +129,7 @@ lint-web:
 lint-backend:
 	cd ai-backend && uv run ruff check src/
 	cd ingestion && uv run ruff check src/
-	cd packages/wahlchat-corpus && uv run ruff check src/
+	cd packages/wahlchat-common && uv run ruff check src/
 
 # --- Testing ---
 
@@ -137,7 +137,7 @@ lint-backend:
 test-backend:
 	cd ai-backend && uv run pytest tests/ --ignore=tests/test_chat_sse.py --ignore=tests/test_local_mode.py
 	cd ingestion && uv run pytest tests/
-	cd packages/wahlchat-corpus && uv run pytest tests/
+	cd packages/wahlchat-common && uv run pytest tests/
 
 test-smoke:
 	cd ai-backend && uv run pytest tests/test_chat_sse.py -x
