+++ ai-backend/pyproject.toml
@@ -11,10 +11,10 @@ license = {text = "PolyForm-Noncommercial-1.0.0"}
 readme = "README.md"
 requires-python = ">=3.12,<3.13"
 
-# Depends on wahlchat-corpus (the shared corpus contract), never on `ingestion` —
+# Depends on wahlchat-common (shared code), never on `ingestion` —
 # that would pull its connector tree (trafilatura, pypdf, bs4, ...) into the image.
 dependencies = [
-    "wahlchat-corpus",
+    "wahlchat-common",
     # HTTP framework + SSE streaming (replaces aiohttp/python-socketio)
     "fastapi>=0.136.3",
     "sse-starlette>=3.4.4",
@@ -41,7 +41,7 @@ dependencies = [
 ]
 
 [tool.uv.sources]
-wahlchat-corpus = { workspace = true }
+wahlchat-common = { workspace = true }
 
 [dependency-groups]
 dev = [
