+++ ai-backend/tests/conftest.py
@@ -370,7 +370,7 @@ def temp_qdrant_collection() -> "Generator[tuple, None, None]":
         from qdrant_client.qdrant_client import QdrantClient as _RealQdrantClient
         from qdrant_client.models import Distance, VectorParams
 
-        from wahlchat_corpus.corpus import EMBEDDING_DIM
+        from wahlchat_common.corpus import EMBEDDING_DIM
 
         client = _RealQdrantClient(url="http://localhost:6333", api_key=None)
         # Quick reachability check — skip fixture if Qdrant is down.
