+++ ai-backend/tests/test_app_rollout_gate.py
@@ -6,7 +6,7 @@ REQUIRE_CORPUS is explicitly set — otherwise local dev / CI boot on empty."""
 import pytest
 
 from src import app as app_module
-from wahlchat_corpus import corpus as setup_collection
+from wahlchat_common import corpus as setup_collection
 
 
 class _FakeCount:
