+++ ingestion/pyproject.toml
@@ -11,11 +11,11 @@ license = {text = "PolyForm-Noncommercial-1.0.0"}
 readme = "README.md"
 requires-python = ">=3.12,<3.13"
 
-# Depends on wahlchat-corpus (the shared corpus contract), never on `ai-backend`.
+# Depends on wahlchat-common (shared code), never on `ai-backend`.
 # Deliberately free of every web-service dependency (fastapi, uvicorn,
 # sse-starlette, firebase-admin) so the Job image carries no chat-service code.
 dependencies = [
-    "wahlchat-corpus",
+    "wahlchat-common",
     "beautifulsoup4>=4.15.0",
     "defusedxml>=0.7.1",
     # Vertex service-account credentials (vertex_credentials.py) and the
@@ -44,7 +44,7 @@ dependencies = [
 ]
 
 [tool.uv.sources]
-wahlchat-corpus = { workspace = true }
+wahlchat-common = { workspace = true }
 
 [dependency-groups]
 dev = [
