+++ ingestion/src/ingestion/connectors/manifesto_uploads/bulk.py
@@ -265,7 +265,7 @@ def ingest(
 
     from qdrant_client import QdrantClient  # noqa: PLC0415
 
-    from wahlchat_corpus.embeddings import get_embeddings  # noqa: PLC0415
+    from wahlchat_common.embeddings import get_embeddings  # noqa: PLC0415
     from ingestion.run import run_connector  # noqa: PLC0415
     from ingestion.setup_collection import (  # noqa: PLC0415
         COLLECTION_NAME,
