+++ ingestion/tests/connectors/abgeordnetenwatch/test_connector.py
@@ -751,7 +751,7 @@ class TestDiscoverFailFast:
         from datetime import date, timedelta
 
         from ingestion.connectors.abgeordnetenwatch import connector as conn_mod
-        from wahlchat_corpus.legislature_config import (
+        from wahlchat_common.legislature_config import (
             LEGISLATURE_CONFIG,
             LegislatureConfig,
         )
@@ -878,7 +878,7 @@ class TestExplicitLegislatureRequired:
         PRIOR Bundestag periods — the same depth policy as the Landtage. A
         default run must not silently stop at one historical period, and it
         must not backfill older periods (those stay explicit-run only)."""
-        from wahlchat_corpus.legislature_config import (
+        from wahlchat_common.legislature_config import (
             FEDERAL_LEGISLATURE_IDS,
             LEGISLATURE_CONFIG,
         )
