+++ packages/wahlchat-common/src/wahlchat_common/governance_levels.py
