+++ packages/wahlchat-common/src/wahlchat_common/vertex_credentials.py
